rename EditableProduct->ProductForEditing. And query/handler
